restructured main.rs

+++ src/utils.rs
@@ -1,5 +1,5 @@
 
-pub fn get_input(day: i32) -> String {
+pub fn get_input(day: u8) -> String {
     let path: String = format!("./input/input{}.txt", day);
     let contents = std::fs::read_to_string(path).expect("Something went wrong reading the file {}");
     contents
